+++ R/Classes.R
@@ -24,10 +24,11 @@ methods::setMethod("initialize",
           "Var",
           function(.Object, name, type, normal) {
             .Object@name <- name
-            type<-as.character(ifelse(type == "numeric" | type == "integer", "numeric",
-                                      ifelse(type == "factor", "factor",
-                                             stop(sprintf("Type unrecognised %s of %s", type, name)))))
-            .Object@type <- type
+            resolved<-resolveType(type, name)
+            if (is.na(resolved))
+              warning(sprintf("Type unrecognised %s of %s: variable will be ignored",
+                              paste(type, collapse = "/"), name))
+            .Object@type <- as.character(resolved)
             .Object@normal <- normal
             return(.Object)
           })
@@ -171,6 +172,8 @@ methods::setMethod("[", "VarGroup", function(x, i) {
     return(x@missing.value)
   else if (i == "missing.value.name")
     return(x@missing.value.name)
+  else if (i %in% c("name", "type", "normal"))
+    return(methods::slot(x, i))
 })
 
 #' Method to access S4 Var elements